+++ example.py
@@ -1,25 +1,51 @@
-from pyconfix import pyconfix, ConfigOption
-import argparse
-import time
+import time, argparse, sys
+from pyconfix import pyconfix, ConfigOption, ConfigOptionType
+import platform
 
 ### This function saves the current configurations in a defconfig-like format.
 ### Custom save functions can be used to export the settings in any format.
-def custom_save(json_data, _):
+def custom_save(json_data, config, is_diff):
     with open("output_defconfig", 'w') as f:
         for key, value in json_data.items():
-            if value == None or (isinstance(value, bool) and value == False):
-                continue
-            if isinstance(value, str):
-                f.write(f"CONFIG_{key}=\"{value}\"\n")
+            if value is None or (isinstance(value, bool) and value == False):
+                f.write(f"# {key} is not set\n")
+            elif isinstance(value, str):
+                f.write(f"{key}=\"{value}\"\n")
             else:
-                f.write(f"CONFIG_{key}={value if value != True else 'y'}\n")
+                f.write(f"{key}={value if value != True else 'y'}\n")
 
 ### This function creates the config object making it accessible for use
 ### by external tools like conan and CMake.
 def create_config():
     config = pyconfix(schem_files=["schem.json"], save_func=custom_save, expanded=True, show_disabled=True)
 
-    ### Actions can be added using a decorator
+    ### Aliases can be registered with a dedicated helper
+    config.register_alias(
+        name='tri-state',
+        option_type=ConfigOptionType.ENUM,
+        choices=["INTEGRATED", "MODULE", "DISABLED"]
+    )
+
+    ### Config options can be added using helpers or plain ConfigOption objects
+    config.add_options(
+        config.option_from_alias('tri-state', name='NEW_MODULE_TRI_STATE_FROM_PYTHON'),
+        config.option_from_alias('string', name='STRING_FROM_ALIAS', default='Default string value'),
+        ConfigOption(
+            name='OS',
+            option_type=ConfigOptionType.STRING,
+            default=platform.system(),
+            external=True
+        ),
+        ConfigOption(
+            name='PYTHON_EVALUATED',
+            option_type=ConfigOptionType.STRING,
+            default=".".join(map(str, sys.version_info[:3])),
+            dependencies=lambda x: x.ENABLE_FEATURE_A,
+            external=True
+        ),
+    )
+
+    ### Actions can also be added using a decorator for ease
     @config.action_option(
         requires=lambda x: x.LOG_LEVEL, 
         dependencies="ENABLE_FEATURE_A",
@@ -30,10 +56,20 @@ def create_config():
         return True
     
     # First define a group
-    deployment_group = config.group_option("deployment", dependencies="ENABLE_FEATURE_A")
+    deployment_group_proxy = config.group_option("deployment", dependencies="ENABLE_FEATURE_A")
 
-    # Then use the group's action_option decorator
-    @deployment_group.action_option(
+    # The group ConfigOption can be accessed using the get() function and
+    # sub options can also be added to the group by directly appending them
+    deployment_group_proxy.get().options.extend([
+        ConfigOption(
+            name='SUB_OPTION',
+            option_type=ConfigOptionType.BOOL,
+            default=False
+        )
+    ])
+
+    # Then use the group's action_option decorator to add actions easier
+    @deployment_group_proxy.action_option(
         requires=lambda x: x.build(),
         dependencies="ENABLE_FEATURE_A"
     )
@@ -50,33 +86,16 @@ def create_config():
         time.sleep(2)
         return True
     
-    ### Config options can also be added by calling extend on the config's options
-    config.options.extend([
-        ConfigOption(
-            name='OS',
-            option_type='string',
-            default="UNIX",
-            external=True
-        ),
-        ConfigOption(
-            name='PYTHON_EVALUATED',
-            option_type='string',
-            default="UNIX",
-            dependencies=lambda x: x.ENABLE_FEATURE_A
-        ),
-    ])
-
     return config
 
-if __name__ == "__main__":
-    #################################################
-    ################ Parse arguments ################
-    #################################################
+def parse_args():
     parser = argparse.ArgumentParser(description="Pyconfix configuration runner")
     parser.add_argument(
         "-l", "--load",
         metavar="FILE",
-        help="Load a configuration file"
+        action="append",
+        default=[],
+        help="Load configuration files"
     )
     parser.add_argument(
         "-r", "--run",
@@ -94,9 +113,14 @@ if __name__ == "__main__":
         help="Run in CLI mode instead of graphical mode"
     )
     parser.add_argument(
-        "-d", "--diff",
+        "--expanded",
         action="store_true",
-        help="Save the setting as diff instead of a full config"
+        help="Default state of groups"
+    )
+    parser.add_argument(
+        "--show-disabled",
+        action="store_true",
+        help="Show disabled options in the interface"
     )
     parser.add_argument(
         "-o", "--option",
@@ -112,6 +136,7 @@ if __name__ == "__main__":
             parser.error(f"Invalid format for option '{item}'. Expected KEY=VALUE.")
         key, value = item.split("=", 1)
         options_dict[key] = value
+
     for key, value in options_dict.items():
         if value.lower() in ["true", "false"]:
             options_dict[key] = value.lower() == "true"
@@ -119,17 +144,25 @@ if __name__ == "__main__":
             options_dict[key] = int(value)
         else:
             options_dict[key] = value
+    args.option = options_dict
+    return args
+
+def main():
+    #################################################
+    ################ Parse arguments ################
+    #################################################
+    args = parse_args()
 
     #################################################
-    ############ Create prconfix instance ###########
+    ############ Create pyconfix instance ###########
     #################################################
     config = create_config()
     
     #################################################
-    ############## Run prconfix examlpe #############
+    ############## Run pyconfix examlpe #############
     #################################################
     ### Config can load files, overlays and run in either TUI or CLI mode
-    config.run(config_file=args.load, overlay=options_dict, graphical=not args.cli, output_diff=args.diff)
+    config.run(config_files=args.load, overlay=args.option, graphical=not args.cli)
 
     ### Option values can be accessed as attributes.
     ### Actions can then be run by calling them as methods.
@@ -141,3 +174,6 @@ if __name__ == "__main__":
             print(f"Trace: {trace}")
         if args.print:
             print(f"{args.print}: {config.get(args.print)}")
+
+if __name__ == "__main__":
+    main()